requirements.txt
aerich==0.5.8; python_version >= "3.7" and python_version < "4.0"
aiosqlite==0.17.0; python_version >= "3.7" and python_version < "4.0"
anyio==3.4.0; python_full_version >= "3.6.2" and python_version >= "3.6"
asgiref==3.4.1; python_version >= "3.6"
asyncpg==0.24.0; python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.6.0"
beautifulsoup4==4.10.0; python_full_version > "3.0.0"
cachetools==5.0.0; python_version >= "3.7" and python_version < "4.0" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
certifi==2021.10.8; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
cffi==1.15.0; python_version >= "3.6"
charset-normalizer==2.0.9; python_full_version >= "3.6.0" and python_version >= "3.6"
click==8.0.3; python_version >= "3.7" and python_version < "4.0"
colorama==0.4.4; python_version >= "3.7" and python_full_version < "3.0.0" and sys_platform == "win32" and platform_system == "Windows" and python_version < "4.0" or sys_platform == "win32" and python_full_version >= "3.5.0" and platform_system == "Windows" and python_version >= "3.7" and python_version < "4.0"
cryptography==35.0.0; python_version >= "3.6"
ddlparse==1.10.0; python_version >= "3.7" and python_version < "4.0"
deprecated==1.2.13; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
dictdiffer==0.9.0; python_version >= "3.7" and python_version < "4.0"
dnspython==2.1.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
ecdsa==0.17.0; python_version >= "2.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
email-validator==1.1.3; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
fastapi==0.68.2; python_full_version >= "3.6.1"
google-api-core==2.6.0; python_version >= "3.6"
google-api-python-client==2.39.0; python_version >= "3.6"
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.0; python_version >= "3.6"
google-auth==2.6.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
google==3.0.0
googleapis-common-protos==1.55.0; python_version >= "3.6"
gunicorn==20.1.0; python_version >= "3.5"
h11==0.12.0; python_version >= "3.6"
hiredis==2.0.0; python_version >= "3.6"
httpcore==0.14.3; python_version >= "3.6"
httplib2==0.20.4; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
httptools==0.2.0
httpx==0.21.1; python_version >= "3.6"
idna==3.3; python_full_version >= "3.6.2" and python_version >= "3.6" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
iso8601==0.1.16; python_version >= "3.7" and python_version < "4.0"
oauth2client==4.1.3
oauthlib==3.2.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
packaging==21.3; python_version >= "3.6"
pika==1.2.0
protobuf==3.19.4; python_version >= "3.6"
pyasn1-modules==0.2.8; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
pyasn1==0.4.8; python_version >= "3.5" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
pycparser==2.21; python_version >= "3.6" and python_full_version < "3.0.0" or python_version >= "3.6" and python_full_version >= "3.4.0"
pydantic==1.8.2; python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.6.1"
pydrive==1.3.1
pyparsing==2.4.7; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.4.0"
pypika-tortoise==0.1.1; python_version >= "3.7" and python_version < "4.0"
python-dotenv==0.19.2; python_version >= "3.5"
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3; python_version >= "3.7" and python_version < "4.0"
pyyaml==6.0; python_version >= "3.6"
redis==4.1.0; python_version >= "3.6"
requests-oauthlib==1.3.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
requests==2.27.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
rfc3986==1.5.0; python_version >= "3.6"
rsa==4.7.2; python_version >= "3.5" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
sentry-sdk==1.4.3
six==1.16.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
sniffio==1.2.0; python_full_version >= "3.6.2" and python_version >= "3.6"
soupsieve==2.3.1; python_version >= "3.6" and python_full_version > "3.0.0"
starlette==0.14.2; python_version >= "3.6" and python_full_version >= "3.6.1"
tortoise-orm==0.17.8; python_version >= "3.7" and python_version < "4.0"
typing-extensions==4.0.0; python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.6.1"
ujson==5.0.0; python_version >= "3.7"
uritemplate==4.1.1; python_version >= "3.6"
urllib3==1.26.7; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.6"
uvicorn==0.15.0
uvloop==0.16.0; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.7"
watchgod==0.7; python_version >= "3.5"
websockets==10.1; python_version >= "3.7"
wrapt==1.13.3; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
